extract chains gas fetching
